+++ src/Products/PloneMeeting/tests/helpers.py
@@ -457,26 +457,30 @@ class PloneMeetingTestingHelpers:
         cfg.at_post_edit_script()
         self.changeUser(currentUser)
 
-    def _enableItemValidationLevels(self, cfg, levels=[]):
-        """Disable every item validation levels."""
+    def _enableItemValidationLevel(self, cfg, level=None, suffix=None):
+        """Enable one or every item validation levels."""
         currentUser = self.member.getId()
         self.changeUser('admin')
         itemValLevels = cfg.getItemWFValidationLevels()
         for itemValLevel in itemValLevels:
-            if not levels or itemValLevel['state'] in levels:
+            if not level or itemValLevel['state'] == level:
                 itemValLevel['enabled'] = '1'
+                if suffix:
+                    itemValLevel['suffix'] = suffix
         cfg.setItemWFValidationLevels(itemValLevels)
         cfg.at_post_edit_script()
         self.changeUser(currentUser)
 
-    def _disableItemValidationLevels(self, cfg, levels=[]):
-        """Disable every item validation levels."""
+    def _disableItemValidationLevel(self, cfg, level=None, suffix=None):
+        """Disable one or every item validation levels."""
         currentUser = self.member.getId()
         self.changeUser('admin')
         itemValLevels = cfg.getItemWFValidationLevels()
         for itemValLevel in itemValLevels:
-            if not levels or itemValLevel['state'] in levels:
+            if not level or itemValLevel['state'] == level:
                 itemValLevel['enabled'] = '0'
+                if suffix:
+                    itemValLevel['suffix'] = suffix
         cfg.setItemWFValidationLevels(itemValLevels)
         cfg.at_post_edit_script()
         self.changeUser(currentUser)