+++ src/Products/PloneMeeting/validators.py
@@ -8,7 +8,9 @@ from collective.contact.plonegroup.utils import get_plone_group_id
 from DateTime import DateTime
 from datetime import date
 from plone import api
+from Products.CMFPlone.utils import safe_unicode
 from Products.PloneMeeting.config import PMMessageFactory as _
+from Products.PloneMeeting.utils import get_item_validation_wf_suffixes
 from Products.PloneMeeting.utils import getInterface
 from Products.validation.interfaces.IValidator import IValidator
 from z3c.form import validator
@@ -162,10 +164,12 @@ class PloneGroupSettingsValidator(validator.SimpleFieldValidator):
 
     def validate(self, value):
         # check that if a suffix is removed, it is not used in MeetingConfig or MeetingItems
-        stored_suffixes = get_all_suffixes()
+        stored_suffixes = get_all_suffixes(only_enabled=True)
         # get removed suffixes...
         saved_suffixes = [func['fct_id'] for func in value]
-        removed_suffixes = list(set(stored_suffixes) - set(saved_suffixes))
+        saved_enabled_suffixes = [func['fct_id'] for func in value if func['enabled']]
+        removed_suffixes = list(set(stored_suffixes) - set(saved_enabled_suffixes))
+        really_removed_suffixes = list(set(stored_suffixes) - set(saved_suffixes))
         removed_plonegroups = [
             get_plone_group_id(org_uid, removed_suffix)
             for org_uid in get_organizations(only_selected=False, the_objects=False)
@@ -193,13 +197,32 @@ class PloneGroupSettingsValidator(validator.SimpleFieldValidator):
                 for org_uid in get_organizations(only_selected=False, the_objects=False):
                     removed_plonegroups.append(get_plone_group_id(org_uid, new_function))
 
-        # check that plone_group not used in MeetingConfigs
+        # check that plonegroups and suffixes not used in MeetingConfigs
         removed_plonegroups = set(removed_plonegroups)
         tool = api.portal.get_tool('portal_plonemeeting')
         for cfg in tool.objectValues('MeetingConfig'):
+            msg = _("can_not_delete_plone_group_meetingconfig",
+                    mapping={'cfg_title': safe_unicode(cfg.Title())})
+            # plonegroups
             if removed_plonegroups.intersection(cfg.getSelectableCopyGroups()):
-                msg = _("can_not_delete_plone_group_meetingconfig",
-                        mapping={'cfg_url': cfg.absolute_url()})
+                raise Invalid(msg)
+            # suffixes, values are like 'suffix_proposing_group_level1reviewers'
+            composed_values_attributes = ['itemAnnexConfidentialVisibleFor',
+                                          'adviceAnnexConfidentialVisibleFor',
+                                          'meetingAnnexConfidentialVisibleFor']
+            for composed_values_attr in composed_values_attributes:
+                values = cfg.getField(composed_values_attr).getAccessor(cfg)()
+                values = [v for v in values
+                          for r in removed_suffixes if r in v]
+                if values:
+                    raise Invalid(msg)
+            # itemWFValidationLevels, may be disabled if validation level also disabled
+            # but not removed
+            item_enabled_val_suffixes = get_item_validation_wf_suffixes(cfg)
+            if set(really_removed_suffixes).intersection(item_enabled_val_suffixes):
+                raise Invalid(msg)
+            all_item_val_suffixes = get_item_validation_wf_suffixes(cfg, only_enabled=False)
+            if set(removed_suffixes).intersection(all_item_val_suffixes):
                 raise Invalid(msg)
         # check that plone_group not used in MeetingItems
         catalog = api.portal.get_tool('portal_catalog')
@@ -213,5 +236,6 @@ class PloneGroupSettingsValidator(validator.SimpleFieldValidator):
                 msg = _(msgid, mapping={'item_url': item.absolute_url()})
                 raise Invalid(msg)
 
-validator.WidgetValidatorDiscriminators(PloneGroupSettingsValidator, field=IContactPlonegroupConfig['functions'])
+validator.WidgetValidatorDiscriminators(
+    PloneGroupSettingsValidator, field=IContactPlonegroupConfig['functions'])
 provideAdapter(PloneGroupSettingsValidator)